+++ project-service/src/main/java/com/yms/projectservice/repository/ProjectRepository.java
@@ -7,13 +7,13 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ProjectRepository extends JpaRepository<Project,Long> {
+public interface ProjectRepository extends JpaRepository<Project,Integer> {
 
     @Query("SELECT p FROM Project p WHERE :memberId MEMBER OF p.teamMemberIds")
     List<Project> findByTeamMemberId(@Param("memberId") Long memberId);
 
     @Query("SELECT p.teamMemberIds FROM Project p WHERE p.id = :projectId")
-    List<Long> findMemberIdsByProjectId(@Param("projectId") Long projectId);
+    List<Integer> findMemberIdsByProjectId(@Param("projectId") Integer projectId);
 
 
 }
